# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GrandGold Marketplace API
  description: |
    GrandGold is a multi-country luxury jewellery marketplace with support for India (INR), UAE (AED), and UK (GBP).
    
    ## Authentication
    Most endpoints require Bearer token authentication. Include the JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - 200 requests per minute for product APIs
    - 100 requests per minute for order/payment APIs
    
    ## Countries
    Supported countries: `IN` (India), `AE` (UAE), `UK` (United Kingdom)
  version: 1.0.0
  contact:
    name: GrandGold API Support
    email: api@thegrandgold.com
  license:
    name: Proprietary
    url: https://thegrandgold.com/terms
servers:
  - url: https://api.thegrandgold.com
    description: Production
  - url: https://staging-api.thegrandgold.com
    description: Staging
  - url: http://localhost:3000
    description: Local Development
...
paths:
...
  /api/products:
    get:
      tags: [Products]
      summary: List products
      operationId: listProducts
      parameters:
        - name: country
          in: query
          schema:
            type: string
            enum: [IN, AE, UK]
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: integer
  /api/products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/orders:
    get:
      tags: [Orders]
      summary: List customer orders
      operationId: listOrders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, shipped, delivered, cancelled]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
  /api/orders/{id}:
    get:
      tags: [Orders]
      summary: Get order details
      operationId: getOrder
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /api/checkout/create:
    post:
      tags: [Orders]
      summary: Create order from cart
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId:
                    type: string
                  paymentRequired:
                    type: boolean

  # ==================== PAYMENTS ====================
...
  /api/payments/razorpay/create-order:
    post:
      tags: [Payments]
      summary: Create Razorpay order
      operationId: createRazorpayOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: integer
                  description: Amount in paise
                currency:
                  type: string
                  default: INR
                receipt:
                  type: string
              required:
                - amount
                - receipt
      responses:
        '200':
          description: Razorpay order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId:
                    type: string
                  amount:
                    type: integer
                  currency:
                    type: string
                  keyId:
                    type: string
  /api/payments/razorpay/verify:
    post:
      tags: [Payments]
      summary: Verify Razorpay payment
      operationId: verifyRazorpayPayment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: string
                paymentId:
                  type: string
                signature:
                  type: string
              required:
                - orderId
                - paymentId
                - signature
      responses:
        '200':
          description: Payment verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
                  paymentId:
                    type: string

  # ==================== FINTECH ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        sku:
          type: string
        price:
          type: number
        currency:
          type: string
        category:
          type: string
        subcategory:
          type: string
        purity:
          type: string
          enum: ['14K', '18K', '22K', '24K']
        weight:
          type: number
        images:
          type: array
          items:
            type: string
        sellerId:
          type: string
        stock:
          type: integer
        isActive:
          type: boolean
        country:
          type: string
...
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        subtotal:
          type: number
        tax:
          type: number
        shipping:
          type: number
        total:
          type: number
        currency:
          type: string
        shippingAddress:
          $ref: '#/components/schemas/Address'
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      properties:
        productId:
          type: string
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
        image:
          type: string
    Address:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
    CreateOrderRequest:
      type: object
      properties:
        cartId:
          type: string
        shippingAddress:
          $ref: '#/components/schemas/Address'
        billingAddress:
          $ref: '#/components/schemas/Address'
        paymentMethod:
          type: string
          enum: [stripe, razorpay, cod]
        deliveryOption:
          type: string
          enum: [standard, express, click_collect]
      required:
        - cartId
        - shippingAddress
        - paymentMethod
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            details:
              type: object
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...